refactor: unit tests that are integration tests
