+++ packages/astro/test/dev-hydration.test.js
@@ -1,6 +1,6 @@
 import * as assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import { createFixture, createRequestAndResponse, runInContainer } from '../test-utils.js';
+import { createFixture, createRequestAndResponse, runInContainer } from './units/test-utils.js';
 
 describe('hydration', () => {
 	it('should not crash when reassigning a hydrated component', {
