+++ packages/astro/test/dev-request-url.test.js
@@ -1,12 +1,12 @@
 import * as assert from 'node:assert/strict';
 import { after, before, describe, it } from 'node:test';
-import { loadFixture } from '../../test-utils.js';
+import { loadFixture } from './test-utils.js';
 
 describe('vite-plugin-astro-server', () => {
 	describe('url', () => {
-		/** @type {import('../../test-utils.js').Fixture} */
+		/** @type {import('./test-utils.js').Fixture} */
 		let fixture;
-		/** @type {import('../../test-utils.js').DevServer} */
+		/** @type {import('./test-utils.js').DevServer} */
 		let devServer;
 
 		before(async () => {
