+++ packages/astro/test/units/integrations/api.test.js
@@ -9,7 +9,6 @@ import {
 	runHookConfigSetup,
 } from '../../../dist/integrations/hooks.js';
 import { defaultLogger } from '../test-utils.js';
-import { loadFixture } from '../../test-utils.js';
 
 const defaultConfig = {
 	root: new URL('./', import.meta.url),
@@ -139,51 +138,6 @@ describe('Integration API', () => {
 		});
 		assert.equal(updatedSettings.config.site, site);
 		assert.equal(updatedSettings.config.integrations.length, 2);
-	});
-
-	describe('Routes resolved hooks', () => {
-		it.skip('should work in dev', {
-			todo: "[p2] Understand why routes aren't deep equal anymore",
-		}, async () => {});
-	});
-
-	describe('Routes setup hook', () => {
-		it('should work in dev', async () => {
-			let routes = [];
-			const fixture = await loadFixture({
-				root: './fixtures/dev-render/',
-				integrations: [
-					{
-						name: 'test',
-						hooks: {
-							'astro:route:setup': (params) => {
-								routes.push({
-									component: params.route.component,
-									prerender: params.route.prerender,
-								});
-							},
-						},
-					},
-				],
-			});
-			const devServer = await fixture.startDevServer();
-
-			try {
-				// The hook should have been called for each route during startup.
-				// Filter to just the project pages we know about.
-				const projectRoutes = routes
-					.filter((r) => r.component.startsWith('src/pages/'))
-					.sort((a, b) => a.component.localeCompare(b.component));
-
-				assert.ok(projectRoutes.length > 0, 'Should have collected routes');
-				// All routes in a static project should be prerendered by default
-				for (const route of projectRoutes) {
-					assert.equal(route.prerender, true, `${route.component} should be prerendered`);
-				}
-			} finally {
-				await devServer.stop();
-			}
-		});
 	});
 });
